Fixed the scope of the useDef links
Prior to this they only showed the question, and not the id and the type. Furthermore, because the scope of the variables is always global, the newly defined variable would overwrite the original value and would be displayed in the default case. Also, this is the reason for the bad scope.